Bunch of fixes to the ui
Still, a lot to do, manageable. 

Updated quite a few things to make it fit within the dark theme. 
Moved some elements around and added a few bits and bobs here and there to make the ui for clean and pc focused. 

Plans - - - - - - - - - - - - - - - - - - - - - - - -
(For the next update)
Add enter button for like.
Allow for arrow key scrolling
Fix the comment section
Fix the hover on the profile icon